fix(rl): divide logits by temperature at fp32, not bf16
Both log-prob paths scaled the bf16 logits before upcasting, so the quotient
was rounded back to bf16. At temperature 0.7 that shifts a log-prob by ~0.13
nats — larger than the rollout-vs-train gap vllm_kl exists to measure. It also
made the local path depend on whether the caller passed already-upcast logits,
so toggling the entropy term (which upcasts) moved the vllm_kl baseline.

Both now scale at fp32; the log-softmax upcast happens anyway, so this only
moves the cast earlier. temperature=1.0 is bit-identical to before, which is
every shipped recipe.

+++ molt/trainer/fsdp/packing.py
@@ -229,7 +229,10 @@ def log_probs_from_vocab_parallel_logits(
         vocab_end = vocab_start + int(shard_sizes[tp_rank].item())
 
     if temperature != 1.0:
-        local_logits = local_logits / temperature
+        # fp32 before the divide, same contract as log_probs_from_logits: rounding the quotient back
+        # to bf16 costs ~1 ULP per logit on top of the logits' own quantization. The log-softmax
+        # below upcasts anyway, so this only moves the cast earlier; autograd casts the grad back.
+        local_logits = local_logits.float() / temperature
     if inference_only is None:
         inference_only = not torch.is_grad_enabled()
     return _DistributedLogProb.apply(local_logits, target, vocab_start, vocab_end, tp_group, inference_only)
